pyproject.toml
[project]
name = "project-sat"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.2",
    "google-cloud>=0.34.0",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-storage>=3.4.0",
    "gunicorn>=23.0.0",
    "polars>=1.33.1",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "ruff>=0.13.0",
    "uvicorn>=0.36.0",
    "typing>=3.10.0.0",
]

